Move component APIs to separate file
Currently the component APIs are grouped into the same file as
manager.go. As more component APIs are added, managers.go will be
difficult to maintain. Similar to the environment and registry APIs,
component APIs will be moved into a separate file component.go
