+++ metadata/manager_test.go
@@ -19,7 +19,6 @@ import (
 	"os"
 	"os/user"
 	"path"
-	"sort"
 	"testing"
 
 	"github.com/spf13/afero"
@@ -205,60 +204,6 @@ func TestFindSuccess(t *testing.T) {
 	f.Close()
 
 	findSuccess(t, appPath, appFile)
-}
-
-func TestComponentPaths(t *testing.T) {
-	spec, err := parseClusterSpec(fmt.Sprintf("file:%s", blankSwagger), testFS)
-	if err != nil {
-		t.Fatalf("Failed to parse cluster spec: %v", err)
-	}
-
-	appPath := AbsPath("/componentPaths")
-	reg := newMockRegistryManager("incubator")
-	m, err := initManager("componentPaths", appPath, spec, &mockAPIServer, &mockNamespace, reg, testFS)
-	if err != nil {
-		t.Fatalf("Failed to init cluster spec: %v", err)
-	}
-
-	// Create empty app file.
-	components := appendToAbsPath(appPath, componentsDir)
-	appFile1 := appendToAbsPath(components, "component1.jsonnet")
-	f1, err := testFS.OpenFile(string(appFile1), os.O_RDONLY|os.O_CREATE, 0777)
-	if err != nil {
-		t.Fatalf("Failed to touch app file '%s'\n%v", appFile1, err)
-	}
-	f1.Close()
-
-	// Create empty file in a nested directory.
-	appSubdir := appendToAbsPath(components, "appSubdir")
-	err = testFS.MkdirAll(string(appSubdir), os.ModePerm)
-	if err != nil {
-		t.Fatalf("Failed to create directory '%s'\n%v", appSubdir, err)
-	}
-	appFile2 := appendToAbsPath(appSubdir, "component2.jsonnet")
-	f2, err := testFS.OpenFile(string(appFile2), os.O_RDONLY|os.O_CREATE, 0777)
-	if err != nil {
-		t.Fatalf("Failed to touch app file '%s'\n%v", appFile1, err)
-	}
-	f2.Close()
-
-	// Create a directory that won't be listed in the call to `ComponentPaths`.
-	unlistedDir := string(appendToAbsPath(components, "doNotListMe"))
-	err = testFS.MkdirAll(unlistedDir, os.ModePerm)
-	if err != nil {
-		t.Fatalf("Failed to create directory '%s'\n%v", unlistedDir, err)
-	}
-
-	paths, err := m.ComponentPaths()
-	if err != nil {
-		t.Fatalf("Failed to find component paths: %v", err)
-	}
-
-	sort.Slice(paths, func(i, j int) bool { return paths[i] < paths[j] })
-
-	if len(paths) != 3 || paths[0] != string(appFile2) || paths[1] != string(appFile1) {
-		t.Fatalf("m.ComponentPaths failed; expected '%s', got '%s'", []string{string(appFile1), string(appFile2)}, paths)
-	}
 }
 
 func TestLibPaths(t *testing.T) {
